Move test function to suite
