+++ lib/domain/TimeLogStatistics.test.ts
@@ -1,14 +1,43 @@
 import TimeLog from "./TimeLog";
 import TimeLogStatistics from "./TimeLogStatistics";
-import { addHours, subDays } from "date-fns";
 import {
-  HOUR,
-  lastMonday,
-  lastSunday,
-  todayWorkdayStart,
-} from "./time-constants";
+  addHours,
+  getDay,
+  nextMonday,
+  nextSunday,
+  subDays,
+  subWeeks,
+} from "date-fns";
+import { HOUR, withTime } from "./time-constants";
 
+const TODAY = new Date();
 export const YESTERDAY = subDays(todayWorkdayStart(), 1);
+
+export function lastMonday() {
+  return lastWeekday(1, nextMonday);
+}
+
+export function lastSunday() {
+  return lastWeekday(0, nextSunday);
+}
+
+export function todayWorkdayStart() {
+  return withTime(new Date(), "09:00:00");
+}
+
+function lastWeekday(dayIndex: number, nextMethod: (date: Date) => Date) {
+  if (getDay(TODAY) === dayIndex) {
+    return subWeeks(TODAY, 1);
+  }
+
+  const day = subWeeks(nextMethod(TODAY), 1);
+
+  if (getDay(day) !== dayIndex) {
+    throw new Error("no correct week day");
+  }
+
+  return day;
+}
 
 function testFulfilledTimeLog(date: Date, hours: number) {
   return new TimeLog({
